LayoutTests/http/tests/site-isolation/select-in-cross-origin-iframe-expected.txt

Test that choosing a select option in a cross-origin iframe updates the element's value when site isolation is enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS selectedValue is "April"
PASS successfullyParsed is true

TEST COMPLETE
